packages/poller/pyproject.toml
[project]
name = "poller"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Etty Sekhon", email = "etty.sekhon@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "curl-cffi>=0.10.0",
    "python-dotenv>=1.1.0",
    "redis>=6.0.0",
    "yfinance>=0.2.59",
]

packages = [
  { include = "poller", from = "src" }
]

[project.scripts]
poller = "poller:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
